Code Changes Against Code review

+++ src/Modules/Execution/Execution.Application/Features/DailyProgress/Commands/DailyProgressCommands.cs
@@ -4,5 +4,5 @@ using Himapp.Execution.Application.Features.DailyProgress.Models;
 namespace Himapp.Execution.Application.Features.DailyProgress.Commands;
 
 public sealed record CreateDailyProgressCommand(CreateDailyProgressRequest Request) : IRequest<DailyProgressModel>;
-public sealed record UpdateDailyProgressCommand(long Id, UpdateDailyProgressRequest Request) : IRequest<DailyProgressModel?>;
-public sealed record DeleteDailyProgressCommand(long Id) : IRequest<bool>;
+public sealed record UpdateDailyProgressCommand(int Id, UpdateDailyProgressRequest Request) : IRequest<DailyProgressModel?>;
+public sealed record DeleteDailyProgressCommand(int Id) : IRequest<bool>;
