+++ src/Modules/Execution/Execution.Application/Features/DailyProgress/Queries/DailyProgressQueries.cs
@@ -4,4 +4,4 @@ using Himapp.Execution.Application.Features.DailyProgress.Models;
 namespace Himapp.Execution.Application.Features.DailyProgress.Queries;
 
 public sealed record GetAllDailyProgressQuery : IRequest<IReadOnlyCollection<DailyProgressModel>>;
-public sealed record GetDailyProgressByIdQuery(long Id) : IRequest<DailyProgressModel?>;
+public sealed record GetDailyProgressByIdQuery(int Id) : IRequest<DailyProgressModel?>;
